refactor: 프로젝트찾을때 userId도 일치하는것으로 찾기

+++ src/routes/controllers/project.controller.js
@@ -64,8 +64,8 @@ const deleteProject = async (req, res, next) => {
       return next(createError(400, BAD_REQUEST));
     }
 
+    await Project.deleteOne({ _id: projectId, owner: userId });
     await User.findByIdAndUpdate(userId, { $pull: { projects: projectId } });
-    await Project.deleteOne({ _id: projectId });
 
     return res.json({ result: SUCCESS });
   } catch (error) {
